Move course field to base notification type
