+++ src/main/webapp/app/entities/group-notification.model.ts
@@ -1,5 +1,4 @@
 import { Notification } from 'app/entities/notification.model';
-import { Course } from 'app/entities/course.model';
 
 export const enum GroupNotificationType {
     INSTRUCTOR = 'INSTRUCTOR',
@@ -8,6 +7,5 @@ export const enum GroupNotificationType {
 }
 
 export class GroupNotification extends Notification {
-    type: GroupNotificationType;
-    course: Course;
+    public type: GroupNotificationType;
 }
